requirements.txt
numpy
pandas
requests
ydata_profiling
dvc
apache-airflow
datasets
dvc[s3]
python-dotenv==0.19.1
dvc-s3
tensorflow-macos
tensorflow-metal
apache-beam
tensorflow-data-validation